requirements.txt
Flask
gunicorn
pandas
numpy
ipykernel
PYyaml
sklearn
-e .